chore: remove tinytest import and place it in suggests
In the current state the package `tinytest2JUnit` does not require the
`tinytest` to be imported.
It does need to be placed in Suggests in the DESCRIPTION file as the
package is needed for to run the tests of the `tinytest2Junit` package.
(We still have to add the test folder! I will make another PR for that).

We should also not import the namespace of tinytest! Even if we do use
functions of tinytests at a later stage with wrappers!
We should then explicitly import the needed functions.

The problem with importing the entire namespace of tinytests is that it
can easily cause namespace clashes.
An @import of a package is really only be done if you use a whole lot
the its functions (like ggplot).

My assumption on why we added the import was tests in tinytests are
written without the `tinytests::` prefix and so one would assume that
the namespace of tinytests needs to be loaded. Indeed tests written with
the tinytest framework do not require the explicit namespace prefix. But
for the tests to run one also does not need to explicitly load the
namespace.

There is actually interesting technical reason on why. If you like to
read more about this:
https://github.com/markvanderloo/tinytest/issues/113

But if one wants to run the test interactivly I do understand it you
want the functions to be loaded. I would say when we write the test
simply in the console first type libary(tinytest).

I hope my explanation was clear :)