fix: update ci code load exported functions of pkg to test
In tinytest one can write tests with the assumption that they are
functions are already loaded. R CMD Check does that for us when it runs
the test. However, it is explicitly stated in the documentation of
tinytest::run_test_dir that it is assumed that all the function are
already loaded.

This commit ensures that the assumption is satisfied. For package that
would need to get tested in the pipeline.

To verify that we are doing it correctly, this commit also removes the
`tinytest2JUnit::` for it's single exported function writeJUnit.