feat: add runTestDir + have info in the testcase tag name.
This commit introduces a new funciton `runTestDir` which requires
explicit presence of the `tinytest` package. It is a drop in replacement
for tinytes::run_test_dir. But with the added benefit that crashing
test_files will get captured and exported as error in the JUnit xml.

This commit also does a large refactoring of the constructTestcaseTag
code. Most notable visible change of end users here is that a testcase
will now have "info" property of the tinytest in the name of the
testcase.

Also error tags are now also reported. Where failing test_files will get
reported as an error-tag in the JUnit. If availalbe, the error
description will next to the error message also show an stacktrace of
where the error occured. With the root of test_file as the hightest
frame in the stack to get reported.