fix: Do not add whitespace to text in xml tag. General formate fix
This commit fixes in general how the xml get's formatted. And makes
the tags stricter.

Mixed tags (content with string and chlid tags) is disallowed.
This is valid xml syntax but does not occur with JUnit so no need to
complicate our own code to support this.

Updates to tag formatting:

* Tags with no content have the closing tag on the same line as the
  opening tag.
* Tags with text contont is placed right inbetween the opening and
  closing tag. Even if the text content has newlines itself.
* Each child tag is placed on a new line with 1 level of indend deeper.