fix: Satisfy JUnit specifcation by adding classname to testcase
Also classname of testcase tags are now the file name (without
file exention as the `.` can incorrectly interpreted as a subclass for
some JUnit parser). Same for 'name' attribute of testsuite tag.

They are also now escaped. For if the file was an illegal xml
Though this would probably not have been the case!

Added also hostname attribute to the testuite tag.

name of testcase now does not include filename info as this info is now
embedded in the class name property.

In addition name of side-effect has been updated to reflect this is this
test was a side effect test.